Sample of mothers and daughters employment histories

A data frame describing mothers employment histories from age 14 to 60 and daughters employment histories from the completion of education to 15 years later. Sequences are sampled (N = 400) from "Biographies et entourage" survey (INED, 2001).

data("seqgimsa")

Format

A data frame with 400 observations and 62 numeric variables. The first 15 variables (prefixed 'f') describe the daughters employment status a given year : 1 = education, 2 = inactivity, 3 = part-time job, 4 = full-time job. The following 47 variables (prefixed 'm') describe the mothers employment status at a given age : 1 = self-employment, 3 = higher level or intermediate occupation, 5 = lower level occupation, 8 = inactivity, 9 = education.

Examples

data(seqgimsa)
str(seqgimsa)
#> 'data.frame':	400 obs. of  62 variables:
#>  $ f1 : num  4 4 4 4 3 4 4 4 4 4 ...
#>  $ f2 : num  4 4 4 4 4 4 4 4 4 4 ...
#>  $ f3 : num  4 4 4 4 4 4 4 4 4 2 ...
#>  $ f4 : num  4 4 4 4 4 4 4 4 4 2 ...
#>  $ f5 : num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f6 : num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f7 : num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f8 : num  4 4 4 2 4 4 4 2 4 2 ...
#>  $ f9 : num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f10: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f11: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f12: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f13: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f14: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ f15: num  4 4 4 4 4 4 4 2 4 2 ...
#>  $ m14: num  5 5 9 9 8 8 9 9 9 9 ...
#>  $ m15: num  5 5 9 9 8 8 9 9 9 9 ...
#>  $ m16: num  5 5 9 9 8 8 9 9 9 9 ...
#>  $ m17: num  5 5 9 9 8 8 9 9 9 9 ...
#>  $ m18: num  5 5 9 3 8 8 9 9 3 5 ...
#>  $ m19: num  5 5 9 3 8 8 9 9 3 5 ...
#>  $ m20: num  5 5 9 3 8 8 9 9 3 5 ...
#>  $ m21: num  5 5 9 3 8 8 9 5 3 5 ...
#>  $ m22: num  5 5 9 3 8 8 9 5 3 5 ...
#>  $ m23: num  5 5 9 3 8 8 9 5 3 5 ...
#>  $ m24: num  5 5 9 3 8 8 9 5 3 8 ...
#>  $ m25: num  5 5 9 3 8 8 1 5 3 8 ...
#>  $ m26: num  5 5 9 3 8 8 1 5 3 8 ...
#>  $ m27: num  5 5 9 3 8 8 1 5 3 8 ...
#>  $ m28: num  5 5 9 3 8 8 1 5 3 8 ...
#>  $ m29: num  5 5 9 3 8 8 1 5 3 8 ...
#>  $ m30: num  5 5 9 8 8 8 1 5 3 8 ...
#>  $ m31: num  5 5 9 8 8 8 1 5 3 8 ...
#>  $ m32: num  5 5 9 8 8 8 1 5 3 8 ...
#>  $ m33: num  5 5 9 8 8 8 1 5 3 8 ...
#>  $ m34: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m35: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m36: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m37: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m38: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m39: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m40: num  5 5 8 8 8 8 1 5 3 8 ...
#>  $ m41: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m42: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m43: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m44: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m45: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m46: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m47: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m48: num  5 5 5 8 8 8 1 5 3 8 ...
#>  $ m49: num  8 5 5 9 8 8 1 5 3 8 ...
#>  $ m50: num  8 5 5 9 8 8 1 5 3 8 ...
#>  $ m51: num  8 5 5 9 8 8 1 5 3 8 ...
#>  $ m52: num  8 5 5 9 8 8 1 5 3 8 ...
#>  $ m53: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m54: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m55: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m56: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m57: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m58: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m59: num  8 5 5 9 8 8 1 5 8 8 ...
#>  $ m60: num  8 8 5 9 8 8 1 8 8 8 ...