Sample of marital, parental and residential sequences

A data frame describing respectively the matrimonial, parental and residential status from age 14 to age 35. It's sampled (N=500) from "Biographies et entourage" survey (INED, 2001).

data("seqmsa")

Format

A data frame with 500 observations and 66 variables. The first 22 variables (prefixed 'log') describe the residential status at a given age : 0 = not independent, 1 = independent. The next 22 variables (prefixed 'mat') describe the matrimonial status at a given age : 1 = never been in a relationship, 2 = cohabiting union, 3 = married, 4 = separated. The last 22 variables (prefixed 'nenf') describe the parental status at a given age : 0 = no child, 1 = one child, 2 = two children, 3 = three children or more.

Examples

data(seqmsa)
str(seqmsa)
#> 'data.frame':	500 obs. of  66 variables:
#>  $ slog14: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ slog15: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ slog16: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ slog17: num  0 0 0 0 0 0 0 0 1 0 ...
#>  $ slog18: num  0 0 0 0 0 0 0 0 1 0 ...
#>  $ slog19: num  0 0 0 0 0 1 0 0 1 0 ...
#>  $ slog20: num  0 0 0 0 1 1 0 0 0 0 ...
#>  $ slog21: num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ slog22: num  0 1 0 1 1 0 0 0 1 0 ...
#>  $ slog23: num  0 1 0 1 1 1 0 1 1 0 ...
#>  $ slog24: num  0 1 0 1 1 1 0 1 1 0 ...
#>  $ slog25: num  0 1 1 1 1 1 0 1 1 0 ...
#>  $ slog26: num  1 1 1 1 1 1 0 1 1 0 ...
#>  $ slog27: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog28: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog29: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog30: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog31: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog32: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog33: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog34: num  1 1 1 1 1 1 0 1 1 1 ...
#>  $ slog35: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat14: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat15: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat16: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat17: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat18: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat19: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ smat20: num  1 3 1 1 3 1 1 1 1 1 ...
#>  $ smat21: num  1 3 1 1 3 3 1 3 1 1 ...
#>  $ smat22: num  1 3 1 3 3 3 1 3 1 1 ...
#>  $ smat23: num  1 3 1 3 3 3 1 3 1 1 ...
#>  $ smat24: num  1 3 1 3 3 3 1 3 1 1 ...
#>  $ smat25: num  1 3 3 3 3 3 1 3 2 1 ...
#>  $ smat26: num  2 3 3 3 3 3 1 3 2 1 ...
#>  $ smat27: num  2 3 3 3 3 3 1 3 2 1 ...
#>  $ smat28: num  2 3 3 3 3 3 1 3 3 1 ...
#>  $ smat29: num  2 3 3 3 3 3 1 3 3 1 ...
#>  $ smat30: num  2 3 3 3 3 3 1 3 3 1 ...
#>  $ smat31: num  3 3 3 3 3 3 1 3 3 1 ...
#>  $ smat32: num  3 3 3 3 2 3 1 3 3 1 ...
#>  $ smat33: num  3 3 3 3 2 3 3 3 3 1 ...
#>  $ smat34: num  3 3 3 3 2 3 3 3 3 1 ...
#>  $ smat35: num  3 3 3 3 2 3 3 3 3 1 ...
#>  $ nenf14: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf15: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf16: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf17: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf18: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf19: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ nenf20: num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ nenf21: num  0 1 0 0 1 0 0 0 0 0 ...
#>  $ nenf22: num  0 1 0 0 1 0 0 0 0 0 ...
#>  $ nenf23: num  0 1 0 0 1 0 0 0 0 0 ...
#>  $ nenf24: num  0 1 0 0 1 1 0 0 0 0 ...
#>  $ nenf25: num  0 1 0 1 1 1 0 1 0 0 ...
#>  $ nenf26: num  0 1 0 1 1 1 0 1 2 0 ...
#>  $ nenf27: num  0 1 0 1 1 1 0 1 2 0 ...
#>  $ nenf28: num  0 1 1 1 1 1 0 2 3 0 ...
#>  $ nenf29: num  0 2 1 2 1 1 0 2 3 0 ...
#>  $ nenf30: num  0 2 2 2 1 1 0 3 3 0 ...
#>  $ nenf31: num  1 2 2 2 1 1 0 3 3 0 ...
#>  $ nenf32: num  1 2 2 2 1 1 0 3 3 0 ...
#>  $ nenf33: num  1 2 2 2 2 1 0 3 3 0 ...
#>  $ nenf34: num  1 2 2 2 2 1 0 3 3 0 ...
#>  $ nenf35: num  1 2 2 2 2 1 0 3 3 0 ...