Sample of employment histories

A data frame describing the employment status from age 14 to age 50. It's a sample of 500 interviewees from "Biographies et entourage" survey (INED, 2001).

data("trajact")

Format

A data frame with 500 observations and 37 variables. Each variable is numeric and describes the employment status at a given age : 1 = education, 2 = full-time job, 3 = part-time job, 4 = small jobs, 5 = inactivity, 6 = military service.

Examples

data(trajact)
str(trajact)
#> 'data.frame':	500 obs. of  37 variables:
#>  $ sact14: num  2 1 1 1 1 1 1 1 1 1 ...
#>  $ sact15: num  2 1 1 1 1 1 1 1 1 1 ...
#>  $ sact16: num  2 1 1 1 1 1 1 1 1 1 ...
#>  $ sact17: num  2 1 1 1 1 1 2 2 2 2 ...
#>  $ sact18: num  2 1 2 2 1 1 2 2 2 2 ...
#>  $ sact19: num  2 3 2 2 1 1 2 2 2 2 ...
#>  $ sact20: num  2 3 2 6 2 2 5 2 6 2 ...
#>  $ sact21: num  2 3 2 2 2 2 5 6 6 2 ...
#>  $ sact22: num  2 3 2 2 2 2 5 6 2 2 ...
#>  $ sact23: num  2 3 2 2 2 2 5 2 2 2 ...
#>  $ sact24: num  2 3 2 2 2 5 3 2 2 2 ...
#>  $ sact25: num  2 3 2 2 2 5 3 2 2 2 ...
#>  $ sact26: num  2 3 2 2 2 5 3 2 2 2 ...
#>  $ sact27: num  2 3 2 2 2 2 3 2 2 2 ...
#>  $ sact28: num  2 3 5 2 2 2 3 2 2 2 ...
#>  $ sact29: num  2 3 2 2 2 2 3 2 2 2 ...
#>  $ sact30: num  2 3 5 2 2 2 3 2 2 2 ...
#>  $ sact31: num  2 3 5 2 2 2 3 2 2 2 ...
#>  $ sact32: num  2 3 5 2 2 2 5 2 2 2 ...
#>  $ sact33: num  2 3 5 2 2 2 5 2 2 2 ...
#>  $ sact34: num  2 3 5 2 2 2 5 2 2 2 ...
#>  $ sact35: num  2 3 5 2 2 2 2 2 2 2 ...
#>  $ sact36: num  2 3 5 2 2 2 2 2 2 2 ...
#>  $ sact37: num  2 3 5 2 2 2 2 2 2 2 ...
#>  $ sact38: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact39: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact40: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact41: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact42: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact43: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact44: num  2 2 5 2 2 2 2 2 2 2 ...
#>  $ sact45: num  2 2 3 2 2 2 2 2 2 2 ...
#>  $ sact46: num  2 2 3 2 2 2 2 2 2 2 ...
#>  $ sact47: num  2 2 3 2 2 2 2 2 2 2 ...
#>  $ sact48: num  2 2 3 2 2 2 2 2 2 2 ...
#>  $ sact49: num  2 2 3 2 2 2 2 2 2 2 ...
#>  $ sact50: num  2 3 3 2 2 2 2 2 2 2 ...